In fixture, remove media profile & perimeter from group v2
